make localcall first arg override handler
